Desktop/EYE/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.3
tensorflow==2.16.1
keras==3.3.3
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
